refactor: rename SpeedTypingScore to SpeedTypingPlayer, and add link between Players and Users
