+++ PatteDoie/Models/Platform/PlatformLobby.cs
@@ -5,7 +5,7 @@
 
         public Guid Id { get; set; }
 
-        public PlatformUser[] Users { get; set; } = [];
+        public List<PlatformUser> Users { get; set; } = [];
 
         public required PlatformUser Creator { get; set; }
