+++ PatteDoie/Models/SpeedTyping/SpeedTypingGame.cs
@@ -3,9 +3,9 @@
     public class SpeedTypingGame
     {
         public Guid Id { get; set; }
-        public SpeedTypingScore[] Scores { get; set; }
-        public SpeedTypingTimeProgress[] TimeProgresses { get; set; }
+        public List<SpeedTypingPlayer> Players { get; set; } = [];
+        public List<SpeedTypingTimeProgress> TimeProgresses { get; set; } = [];
         public DateTime LaunchTime { get; set; }
-        public string[] Words { get; set; }
+        public List<string> Words { get; set; } = [];
     }
 }
