--- PatteDoie/Models/SpeedTyping/SpeedTypingScore.cs
@@ -1,9 +1,0 @@
-﻿namespace PatteDoie.Models.SpeedTyping
-{
-    public class SpeedTypingScore
-    {
-        public Guid Id { get; set; }
-        public Guid UserId { get; set; }
-        public int Score { get; set; }
-    }
-}
