+++ PatteDoie/PatteDoieContext.cs
@@ -23,7 +23,7 @@ public partial class PatteDoieContext : DbContext
 
     public DbSet<PatteDoie.Models.SpeedTyping.SpeedTypingTimeProgress> SpeedTypingTimeProgress { get; set; } = default!;
 
-    public DbSet<PatteDoie.Models.SpeedTyping.SpeedTypingScore> SpeedTypingScore { get; set; } = default!;
+    public DbSet<PatteDoie.Models.SpeedTyping.SpeedTypingPlayer> SpeedTypingPlayer { get; set; } = default!;
 
     public DbSet<PatteDoie.Models.SpeedTyping.SpeedTypingGame> SpeedTypingGame { get; set; } = default!;
