+++ PatteDoie/Services/SpeedTyping/ISpeedTypingService.cs
@@ -14,8 +14,8 @@ namespace PatteDoie.Services.SpeedTyping
 
         public Task UpdateGame(Guid id, CreateSpeedTypingGameCommand game);
 
-        public Task<SpeedTypingGameRow> CreateGame(CreateSpeedTypingGameCommand command, PlatformUser[] platformUsers);
-
+        public Task<SpeedTypingGameRow> CreateGame(CreateSpeedTypingGameCommand command, List<PlatformUser> platformUsers);
+            
         public Task DeleteGame(Guid id);
     }
 }
